requirements.txt
# Core numerical stack
numpy
pandas

# ML libraries
scikit-learn
torch

# Plotting
matplotlib

# Utilities
PyYAML
loguru


seaborn
 